+++ demos/sample/src/application.ts
@@ -3,7 +3,10 @@ import {
   canvasText,
   createContext,
   createFontLoader,
+  createInput,
   createTextureLoader,
+  createTimer,
+  inputControlledCamera,
   pipeline,
   sprite,
   tilingSprite,
@@ -106,36 +109,13 @@ export async function application(canvas: HTMLCanvasElement): Promise<void> {
   text.x = canvas.width / 2 - text.width / 2;
   text.y = 250;
 
-  let cameraX = 0;
-  let cameraY = 0;
+  const input = createInput();
+  const timer = createTimer();
+  const camera = inputControlledCamera(input, timer, context);
 
-  document.addEventListener('keydown', (event) => {
-    let x = 0;
-    let y = 0;
-
-    if (['ArrowRight', 'KeyD'].includes(event.code)) {
-      x += 10;
-    }
-
-    if (['ArrowLeft', 'KeyA'].includes(event.code)) {
-      x -= 10;
-    }
-
-    if (['ArrowUp', 'KeyW'].includes(event.code)) {
-      y -= 10;
-    }
-
-    if (['ArrowDown', 'KeyS'].includes(event.code)) {
-      y += 10;
-    }
-
-    cameraX += x;
-    cameraY += y;
-
-    context.observe([cameraX, cameraY]);
-  });
-
-  const draw = function () {
+  const draw = function (now: number) {
+    timer.update(now);
+    camera.update();
     stats.begin();
 
     for (const sprite of rotatingSprites) {
@@ -148,5 +128,5 @@ export async function application(canvas: HTMLCanvasElement): Promise<void> {
     requestAnimationFrame(draw);
   };
 
-  draw();
+  draw(performance.now());
 }